projets/Archives/btext_cleaning_pipeline-lire-le-txt/requirements.txt
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.3.0
tqdm==4.66.1
pydantic==1.10.12